fix: Diff-Review aufgeloest — zwei echte Befunde (AGE-903)
opencode (Kimi-K3) und gemini auf 1944 Diffzeilen, beide REQUEST-CHANGES.
Sechs Befunde, jeder nachgemessen.

Uebernommen:

A · Die it.each-Liste in AppShell.identity.test.tsx deckte FUENF Stufen ab und
    behauptete sechs: `boost` fehlte, `discover` stand doppelt. Ursache ist
    meine mechanische Umbenennung "exchange" -> "discover" — an dieser Stelle
    stand `discover` schon. Die Suche danach fand eine ZWEITE Kollision, die
    kein Reviewer hatte: MitgliedschaftPage.test.tsx:144, derselbe Fehler. Das
    ist der eigentliche Ertrag: ein mechanischer Ersatz kollidiert dort, wo das
    Ziel schon vorkam, und sieht danach vollkommen plausibel aus.

B · MemberDirectory blendete die vier Filter waehrend des Ladens auch fuer
    Clubmitglieder aus — `(levelRank ?? 0)` machte aus "noch nicht geladen" ein
    "Rang 0". Jetzt `levelRank === null || levelRank >= CLUB_RANK`, dieselbe
    Regel wie in MembershipGate und HeaderSearch. Zusage RED gefahren.

C · Der Schluss-Waechter der Migration castet jetzt `::int`. Der Befund selbst
    ist gegenstandslos (`price_year` IST integer, gemessen), aber ohne Cast
    haengt eine Zusage an einer Messung. Beide Richtungen gefahren: schweigt
    bei der richtigen Leiter, bricht bei discover = 299 EUR.

Nicht uebernommen, mit Messung:

D · geminis HOCH-Befund "die Preise fuer focus und impact fehlen" ist falsch
    und hat die verkehrte Ausfallrichtung: beide trugen ihre Preise vorher und
    tragen sie unveraendert (gemessen 5/600 und 6/1200).

E · opencodes Einwand, der Diff belege keine weiteren Aufrufer von
    darfKontaktanfrageSenden, ist methodisch richtig — Messung nachgereicht:
    ein Aufruf, tsc sauber.

Der lokale Stack ist jetzt migriert (Freigabe der Parallelsitzung eingeholt).
40 pgTAP-Dateien mit 1391 Zusagen ueber `supabase test db` gruen, darunter
grants_test. 2928 Vitest-Zusagen, tsc sauber, 15 Deno-Zusagen.

+++ supabase/migrations/20260926120000_stufen_v5.sql
@@ -476,11 +476,15 @@ comment on table public.membership_tiers is
 -- Preis: beides sind gültige Werte in gültigen Zeilen. Genau dafür ist diese
 -- Prüfung da. Gegenprobe ebenfalls gefahren: mit `boost` auf Rang 9 bricht sie
 -- ab und schreibt die vorgefundene Leiter in die Meldung.
+-- Die Zahlen werden vor der Verkettung ausdruecklich auf `int` gecastet. Sie
+-- SIND `integer` (gemessen), der Vergleich stimmt also auch ohne — aber dann
+-- haengt eine Zusage an einer Textdarstellung, und ein spaeterer Leser muesste
+-- den Spaltentyp nachschlagen, um ihr zu trauen.
 do $$
 declare
   v_leiter text;
 begin
-  select string_agg(key || '=' || level_rank || '/' || price_year || '€', ' '
+  select string_agg(key || '=' || level_rank::int || '/' || price_year::int || '€', ' '
                     order by level_rank)
     into v_leiter from public.membership_tiers;
   if v_leiter is distinct from
